+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+# This is a generated file; do not edit or check into version control.
+export "FLUTTER_ROOT=/Users/shubhamkumar/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/shubhamkumar/Sites/projects/flutter-carousel/example"
+export "FLUTTER_TARGET=/Users/shubhamkumar/Sites/projects/flutter-carousel/example/lib/main.dart"
+export "FLUTTER_BUILD_DIR=build"
+export "SYMROOT=${SOURCE_ROOT}/../build/ios"
+export "FLUTTER_FRAMEWORK_DIR=/Users/shubhamkumar/flutter/bin/cache/artifacts/engine/ios"
+export "FLUTTER_BUILD_NAME=1.0.0"
+export "FLUTTER_BUILD_NUMBER=1"