+++ lib/carousel.dart
@@ -1,3 +1,4 @@
 library flutter_multi_carousel;
 
 export 'src/carousel/carousel.dart';
+export 'src/services/type_declaration.dart';