+++ lib/src/carousel/slide_swipe.dart
@@ -1,20 +1,21 @@
 import 'dart:math';
 import 'package:flutter/material.dart';
 
+import '../../carousel.dart';
+
 class SlideSwipe extends StatelessWidget {
   int currentPage;
   bool initial;
-  final dynamic props;
-  Function updatePosition;
+  final Carousel props;
 
-  SlideSwipe(
-    this.props, this.updatePosition
-  );
+  SlideSwipe(this.props);
 
   initiate(index) {
     try {
-      currentPage = props.controller.page.round();
-    } catch (e) {}
+      currentPage = props.controller.initialPage.round();
+    } catch (e) {
+      print("exception here => $e");
+    }
     double value;
     if (index == currentPage - 1 && initial) value = 1.0;
     if (index == currentPage && initial) value = 0.0;
@@ -35,7 +36,7 @@ class SlideSwipe extends StatelessWidget {
         controller: props.controller,
         itemCount: count,
         onPageChanged: (i) {
-          updatePosition(i);
+          props.updatePositionCallBack(i);
           if (props.onPageChange != null) {
             props.onPageChange(i);
           }