+++ lib/src/services/type_declaration.dart
@@ -1,0 +1,9 @@
+enum Types {
+  simple,
+  slideSwiper,
+  xRotating,
+  yRotating,
+  zRotating,
+  multiRotating
+}
+enum IndicatorTypes { bar, dot, bubble }