+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-## [0.0.1] - TODO: Add release date.
+## [1.0.1] - 10 December 2017
 
-* TODO: Describe initial release.
+* Initial release.
+
+## [1.0.2] - 26 February 2019
+* Added feature to initialise the starting page index
+* Added a callback function on tap
+* Added wrap control to carousel
+* Indroduction of Types to select the type of carousel
+* Indroduction of IndicatorTypes to select the type of Indicator